+++ CSM Database Core/CHANGELOG.md
@@ -14,6 +14,20 @@
 | Microsoft.EntityFrameworkCore           | 9.0.8            | 9.0.8           |
 | Microsoft.EntityFrameworkCore.SqlServer | 9.0.8            | 9.0.8           |
 
+## [3.0.0] - 26.11-2025
+
+### Changed
+
+- Now DepotBase operations are based on interfaces trading instead of object classes.
+
+#### Dependencies
+
+| Package                                 | Previous Version | New Version     |
+|:----------------------------------------|:----------------:|:---------------:|
+| CSM.Foundation.Core                     | 2.0.0            | 2.0.0           |
+| Microsoft.EntityFrameworkCore           | 9.0.8            | 9.0.8           |
+| Microsoft.EntityFrameworkCore.SqlServer | 9.0.8            | 9.0.8           |
+
 ## [2.1.0] - 26.11-2025
 
 ### Changed